@@ -24,7 +24,6 @@ import com.kronotop.server.resp3.SimpleStringRedisMessage;
 import com.kronotop.sql.ExecutionContext;
 import com.kronotop.sql.SqlExecutionException;
 import com.kronotop.sql.SqlService;
-import com.kronotop.sql.TransactionResult;
 import com.kronotop.sql.backend.ddl.ColumnAlreadyExistsException;
 import com.kronotop.sql.backend.ddl.model.ColumnModel;
 import com.kronotop.sql.backend.ddl.model.TableModel;
@@ -34,6 +33,7 @@ import org.apache.calcite.sql.ddl.SqlColumnDeclaration;
 import org.apache.calcite.sql.type.SqlTypeName;
 
 import java.util.LinkedHashMap;
+import java.util.concurrent.CompletableFuture;
 
 /**
  * AddColumn is a class that represents the ALTER TABLE operation to add a column to a table in the database.
@@ -90,7 +90,7 @@ public class AddColumn extends BaseAlterType implements AlterType {
         return new SimpleStringRedisMessage(Response.OK);
     }
 
-    public void notifyCluster(TransactionResult result, ExecutionContext context, SqlAlterTable sqlAlterTable) {
-        publishTableAlteredEvent(result, context, sqlAlterTable);
+    public void notifyCluster(CompletableFuture<byte[]> versionstampFuture, ExecutionContext context, SqlAlterTable sqlAlterTable) {
+        publishTableAlteredEvent(versionstampFuture, context, sqlAlterTable);
     }
 }
