@@ -24,12 +24,11 @@ import com.kronotop.server.resp3.SimpleStringRedisMessage;
 import com.kronotop.sql.ExecutionContext;
 import com.kronotop.sql.SqlExecutionException;
 import com.kronotop.sql.SqlService;
-import com.kronotop.sql.TransactionResult;
 import com.kronotop.sql.backend.ddl.model.ColumnModel;
 import com.kronotop.sql.backend.ddl.model.TableModel;
 import com.kronotop.sql.parser.SqlAlterTable;
 
-import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
 public class RenameColumn extends BaseAlterType implements AlterType {
 
@@ -53,7 +52,7 @@ public class RenameColumn extends BaseAlterType implements AlterType {
     }
 
     @Override
-    public void notifyCluster(TransactionResult result, ExecutionContext context, SqlAlterTable sqlAlterTable) {
-        publishTableAlteredEvent(result, context, sqlAlterTable);
+    public void notifyCluster(CompletableFuture<byte[]> versionstampFuture, ExecutionContext context, SqlAlterTable sqlAlterTable) {
+        publishTableAlteredEvent(versionstampFuture, context, sqlAlterTable);
     }
 }
