@@ -21,28 +21,28 @@ import com.kronotop.server.Response;
 import com.kronotop.server.resp3.ErrorRedisMessage;
 import com.kronotop.server.resp3.SimpleStringRedisMessage;
 import com.kronotop.sql.BaseHandlerTest;
+import com.kronotop.sql.backend.AssertResponse;
 import io.lettuce.core.codec.StringCodec;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 public class CreateTableTest extends BaseHandlerTest {
+
     @Test
-    public void test_Success() {
+    public void test_success() {
         KronotopCommandBuilder<String, String> cmd = new KronotopCommandBuilder<>(StringCodec.ASCII);
 
         ByteBuf buf = Unpooled.buffer();
         cmd.sql("CREATE TABLE public.users (id INTEGER, username VARCHAR)").encode(buf);
-        cmd.sql("CREATE TABLE myapp.deployment-1.users (id INTEGER, username VARCHAR)").encode(buf);
         channel.writeInbound(buf);
         Object response = channel.readOutbound();
 
-        assertInstanceOf(SimpleStringRedisMessage.class, response);
-        SimpleStringRedisMessage actualMessage = (SimpleStringRedisMessage) response;
-        assertEquals(Response.OK, actualMessage.content());
+        AssertResponse<SimpleStringRedisMessage> assertResponse = new AssertResponse<>();
+        SimpleStringRedisMessage message = assertResponse.getMessage(response, 0, 1);
+        assertEquals(Response.OK, message.content());
     }
 
     @Test
@@ -56,9 +56,9 @@ public class CreateTableTest extends BaseHandlerTest {
             channel.writeInbound(buf);
             Object response = channel.readOutbound();
 
-            assertInstanceOf(SimpleStringRedisMessage.class, response);
-            SimpleStringRedisMessage actualMessage = (SimpleStringRedisMessage) response;
-            assertEquals(Response.OK, actualMessage.content());
+            AssertResponse<SimpleStringRedisMessage> assertResponse = new AssertResponse<>();
+            SimpleStringRedisMessage message = assertResponse.getMessage(response, 0, 1);
+            assertEquals(Response.OK, message.content());
         }
 
         {
@@ -67,9 +67,10 @@ public class CreateTableTest extends BaseHandlerTest {
             channel.writeInbound(buf);
             Object response = channel.readOutbound();
 
-            assertInstanceOf(ErrorRedisMessage.class, response);
-            ErrorRedisMessage actualMessage = (ErrorRedisMessage) response;
-            assertEquals("SQL Table 'users' already exists", actualMessage.content());
+            AssertResponse<ErrorRedisMessage> assertResponse = new AssertResponse<>();
+            ErrorRedisMessage message = assertResponse.getMessage(response, 0, 1);
+            assertEquals("SQL Table 'users' already exists", message.content());
+
         }
     }
 }
